ulong

+++ src/Silk.Shared/Configuration/SilkConfigurationOptions.cs
@@ -45,9 +45,9 @@ namespace Silk.Shared.Configuration
         public string LogLevel { get; set; } = "Info";
 
         /// <summary>
-        /// Property for holding GuildId's for configuring which Guilds you want to add Music commands to
+        /// Property for holding GuildId's for configuring which Guilds you want to add Music commands to.
         /// </summary>
-        public List<int> PrivateMusic { get; set; }
+        public List<ulong> PrivateMusic { get; set; }
 
         /// <summary>
         /// Property for holding Persistence options (property name matching sub-key property in configuration file)
